refactor code, use Navbar, SideBar and Footer to wrap everything in App.js, resolve rendering error while dispatching isUser state change, implemented the isUser state change in App.js, and dispatched it in login action creator
