+++ src/App.js
@@ -4,32 +4,50 @@ import PetListPage from "./pages/PetListPage";
 import { RegisterPage } from "./pages/RegisterPage";
 import { AddPet } from "./pages/AddPet";
 import PetManagePage from "./pages/PetManagePage";
-import { Suspense } from "react";
+import { Suspense, useEffect, useState } from "react";
 import Spinner from "./components/Spinner";
 import Foundations from "./pages/Foundations";
 import Home from "./pages/Home";
-import { Provider } from "react-redux";
-import { store } from "./store/store";
 import history from "./history";
+import SideBar from "./components/SideBar";
+import Navbar from "./components/Navbar";
+import Footer from "./components/Footer";
+import { useDispatch } from "react-redux";
+import { ISUSER } from "./store/actions";
 
 function App() {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const toggle = () => {
+    setIsOpen(!isOpen);
+  };
+
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    if (localStorage.getItem("Authorization")) {
+      dispatch({ type: ISUSER, payload: true });
+    }
+  }, [dispatch]);
+
   return (
-    <Provider store={store}>
-      <Router history={history}>
-        <Suspense fallback={<Spinner />}>
-          <Switch>
-            <Route exact path="/" component={Home} />
-            <Route exact path="/login" component={LoginPage} />
-            <Route exact path="/signup" component={RegisterPage} />
-            <Route exact path="/foundations/:id/pets" component={PetListPage} />
-            <Route exact path="/pets/:id/request" />
-            <Route exact path="/foundations/:id/add-pet" component={AddPet} />
-            <Route exact path="/pets/:id/manage" component={PetManagePage} />
-            <Route exact path="/foundations" component={Foundations} />
-          </Switch>
-        </Suspense>
-      </Router>
-    </Provider>
+    <Router history={history}>
+      <Suspense fallback={<Spinner />}>
+        <SideBar isOpen={isOpen} toggle={toggle} />
+        <Navbar toggle={toggle} />
+        <Switch>
+          <Route exact path="/" component={Home} />
+          <Route exact path="/login" component={LoginPage} />
+          <Route exact path="/signup" component={RegisterPage} />
+          <Route exact path="/foundations/:id/pets" component={PetListPage} />
+          <Route exact path="/pets/:id/request" />
+          <Route exact path="/foundations/:id/add-pet" component={AddPet} />
+          <Route exact path="/pets/:id/manage" component={PetManagePage} />
+          <Route exact path="/foundations" component={Foundations} />
+        </Switch>
+        <Footer />
+      </Suspense>
+    </Router>
   );
 }
